+++ src/authorization.ts
@@ -1,20 +1,27 @@
+import { IOrder } from './order';
+import { IGood, IOwnershipInfo } from './ownershipInfo';
 import { IProject } from './project';
 import { SortType } from './sortType';
 
+// strict definetion(2023-01-31~)
+export type IOrderAsObject = Pick<IOrder, 'orderNumber' | 'typeOf'>;
+// export type IObject = IOwnershipInfo<IGood> | ISimpleOrder;
+export type IObject = IOwnershipInfo<IGood> | IOrderAsObject;
+
 /**
- * 承認インターフェース
+ * 承認
  */
 export interface IAuthorization {
     project: Pick<IProject, 'id' | 'typeOf'>;
     typeOf: 'Authorization';
     code: string;
-    object: any;
+    object: IObject;
     validFrom: Date;
     validUntil: Date;
 }
 
 /**
- * ソート条件インターフェース
+ * ソート条件
  */
 export interface ISortOrder {
     validFrom?: SortType;
@@ -30,7 +37,7 @@ export interface IObjectSearchConditions {
 }
 
 /**
- * 承認検索条件インターフェース
+ * 承認検索条件
  */
 export interface ISearchConditions {
     limit?: number;