+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - 予約使用アクション最適化
 - 承認オブジェクトとしての注文を最適化
+- 承認オブジェクトとしての所有権を最適化
 
 ### Deprecated