+++ CHANGELOG.md
@@ -18,6 +18,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.284.0 - 2023-02-01
+
+### Added
+
+- 単価オファーにsettingsを追加
+
+### Changed
+
+- 予約使用アクション最適化
+- 承認オブジェクトとしての注文を最適化
+- 承認オブジェクトとしての所有権を最適化
+- Permit所有権のtypeOfGoodを最適化
+
+### Removed
+
+- プロジェクトのonReservationUsed設定を廃止
+
 ## v4.283.0 - 2023-01-18
 
 ### Changed